% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_deployment.R
\name{amlr_deployment}
\alias{amlr_deployment}
\title{Extract deployment information}
\usage{
amlr_deployment(file.l1, n = 1, na.start.skip = TRUE)
}
\arguments{
\item{file.l1}{file path for AMLR L1 nc file}

\item{n}{numeric; number of lat/lon positions to print. Default is 1}

\item{na.start.skip}{logical; indicates if \code{NA} lat/lon data should be
removed from the front of the lat/lon data.
Default is \code{TRUE}}
}
\value{
A named list of 1) the minimum value for the 'time' variable (with \code{NA}s removed) and
2) a data frame with 'lat' and 'lon' columns that contains the first \code{n} rows
of latitude and longitude values, respectively, from \code{file.l1.}.
The names of these elements are 'deployment_time' and 'deployment_position', respectively.
If \code{na.start.skip} is \code{TRUE}, then front rows
with \code{NA} lat or lon values are removed before returning the first \code{n} rows.
See Details for a description of 'front'

A warning is printed if any of the printed coordinates are not
south of 58 deg S and between 70 deg W and 50 deg W.
}
\description{
Extract deployment time and location from AERD L1 data
}
\details{
The purpose of this function is to extract and print the deployment time and
position for \code{file.l1}, which is expected to be an AMLR L1 file.
This deployment info is 1) the minimum value in the 'time' variable and
2) the first one or more recorded positions to determine the approximate deployment location.
For positions, this function extracts data for the variables 'latitude' and 'longitude'.
For this function, 'front' (e.g. front rows with \code{NA} lat/lon data)
means before the first record with non-\code{NA} lat and lon data.
}
