% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_deployment_position.R
\name{amlr_deployment_position}
\alias{amlr_deployment_position}
\title{Extract deployment location}
\usage{
amlr_deployment_position(file.l1, n = 1, na.start.skip = TRUE)
}
\arguments{
\item{file.l1}{file path for AMLR L1 nc file}

\item{n}{numeric; number of lat/lon positions to print. Default is 1}

\item{na.start.skip}{logical; indicates if \code{NA} lat/lon data should be
removed from the front of the lat/lon data.
Default is \code{TRUE}}
}
\value{
Data frame with 'lat' and 'lon' columns that contains the first \code{n} rows
of latitude and longitude values, respecitvely, from \code{file.l1.}.
If \code{na.start.skip} is \code{TRUE}, then front rows
with \code{NA} lat or lon values are removed before returning the first \code{n} rows.
See Details for a description of 'front'

A warning is printed if any of the printed coordiantes are not
south of 58 deg S and between 70 deg W and 50 deg W.
}
\description{
Extract deployment location from AERD L1 data
}
\details{
The purpose of this function is to extract the first one or more
recorded positions to determine the approximate deployment location.
This function extracts data for the variables 'latitude' and 'longitude'
from \code{file.l1}. \code{file.l1} is expected to be an AMLR L1 file.
For this function, 'front' (e.g. front rows with \code{NA} lat/lon data)
means before the first record with non-\code{NA} lat and lon data.
}
