% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_ngdac_nc_put.R
\name{amlr_ngdac_nc_put}
\alias{amlr_ngdac_nc_put}
\title{Add values and attributes to nc file}
\usage{
amlr_ngdac_nc_put(
  ncnew,
  profile.curr,
  ts.curr,
  glider.name,
  wmo.id,
  ctd.calib.date,
  ctd.calib.date.factory,
  ctd.serial.num
)
}
\arguments{
\item{ncnew}{nc file to which to add data and attributes; output of \code{nc_create}}

\item{profile.curr}{list of data extracted from AERD L2 file.
This list must contain the following named components: TODO}

\item{ts.curr}{list of data extracted from AERD L! file
This list must contain the following named components: TODO}

\item{glider.name}{character; name of glider, e.g. "amlr01"}

\item{wmo.id}{character; the WMO ID of the glider/deployment}

\item{ctd.calib.date}{Date; date of CTD instrument calibration}

\item{ctd.calib.date.factory}{Date; date of CTD instrument factory calibration}

\item{ctd.serial.num}{character; serial number of CTD instrument}
}
\value{
\code{TRUE} if all code is executed without errors
}
\description{
Add variable values and both variable and global attributes to nc file
}
\details{
This function should only be called by \code{\link{amlr_ngdac_convert}}.
It adds attributes, and data as applicable, both globally and to defined variables
for the nc file \code{ncnew}.

Dynamic (aka deployment-specific) information is either calculated from
the data (e.g. deployment date/time)
or passed to function via arguments (e.g. CTD calibration information)
}
