% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlrGlider-package.R
\docType{package}
\name{swfscAirDAS-package}
\alias{swfscAirDAS-package}
\alias{swfscAirDAS}
\title{AMLR Glider Data Processing}
\description{
Process AMLR glider data
}
\details{
This package contains functions designed for processing AMLR glider data.
Currently this means converting AMLR glider output nc files to the nc
file format required by IOOS NGDAC
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
