% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amlr_ngdac_convert.R
\name{amlr_ngdac_convert}
\alias{amlr_ngdac_convert}
\title{Convert AMLR L1 and L2 glider data to NGDAC format v2.0}
\usage{
amlr_ngdac_convert(file.l1, file.l2, path.out, glider.name, ...)
}
\arguments{
\item{file.l1}{file path for AMLR L1 nc file}

\item{file.l2}{file path for AMLR L2 nc file}

\item{path.out}{file path to folder where output nc files will be created}

\item{glider.name}{character; name of glider}

\item{...}{arguments passed to \code{\link{amlr_ngdac_nc_put}}}
}
\value{
\code{TRUE} if the function ran successfully.
This functions writes one nc file per profile to the folder specified by \code{path.out}
}
\description{
Convert AMLR L1 and L2 glider data to NGDAC format v2.0
}
\details{
This is the top-level function for converting AMLR glider data to NGDAC format v2.0,
as described at the url below. The AMLR glider data is the output from
\url{https://github.com/socib/glider_toolbox}, and thus consists of
an L1 and L2 file containing time series (trajectory) and profile data, respectively.
Data from both of these files is used to create the NGDAC-formatted nc files.

In addition to the variables time series variables listed at the NGDAC website,
this function writes the following variables to the output nc files:
TODO
}
\examples{

\dontrun{
amlr_ngdac_convert(
  file.l1 = "AERD_18_19/dep100_amlr01_sgg3_L1_2018-12-06_data_rt.nc",
  file.l2 = "AERD_18_19/dep100_amlr01_sgg3_L2_2018-12-06_data_rt.nc",
  file.out.path = "Output/dep100_amlr01_sgg3/",
  glider.name = "amlr01"
)
}

}
\seealso{
\url{https://ioos.github.io/ioosngdac/ngdac-netcdf-file-format-version-2.html}

\url{http://cfconventions.org/Data/cf-standard-names/73/build/cf-standard-name-table.html}
}
